:recycle: Refactor test structure
React components and their tests are kept together in the same directory
for easier finding.
