+++ src/FormStart/tests.spec.js
@@ -1,17 +1,19 @@
 import React from 'react';
 import { render, unmountComponentAtNode } from "react-dom";
 import { act } from "react-dom/test-utils";
-import FormStart from "../../../FormStart";
-import {testForm} from "./fixtures/form";
-import useQuery from "../../../hooks/useQuery";
 
-jest.mock("../../../hooks/useQuery");
+import useQuery from '../hooks/useQuery';
+import FormStart from '.';
+
+import {testForm} from './fixtures';
+
+jest.mock('../hooks/useQuery');
 
 let container = null;
 
 beforeEach(() => {
   // setup a DOM element as a render target
-  container = document.createElement("div");
+  container = document.createElement('div');
   document.body.appendChild(container);
 });
 
@@ -22,8 +24,8 @@ afterEach(() => {
   container = null;
 });
 
-it("Form start page displays DigiD generic error", () => {
-  const testLocation = new URLSearchParams("?_start=1&_digid-message=error");
+it('Form start page displays DigiD generic error', () => {
+  const testLocation = new URLSearchParams('?_start=1&_digid-message=error');
   useQuery.mockReturnValue(testLocation);
 
   act(() => {
@@ -39,8 +41,8 @@ it("Form start page displays DigiD generic error", () => {
   expect(container.textContent).toContain('Er is een fout opgetreden bij het inloggen met DigiD. Probeer het later opnieuw.');
 });
 
-it("Form start page displays DigiD cancel login error", () => {
-  const testLocation = new URLSearchParams("?_start=1&_digid-message=login-cancelled");
+it('Form start page displays DigiD cancel login error', () => {
+  const testLocation = new URLSearchParams('?_start=1&_digid-message=login-cancelled');
   useQuery.mockReturnValue(testLocation);
 
   act(() => {
