--- src/jstests/dummy.spec.js
@@ -1,6 +1,0 @@
-import React from 'react';
-
-test('First test', () => {
-    const a = 1;
-    expect(a).toEqual(1);
-});
